requirements.txt
Flask==2.3.2
streamlit==1.23.1
matplotlib==3.7.2
scikit-learn==1.2.2
numpy==1.24.2
requests==2.31.0
beautifulsoup4==4.11.1
sentence-transformers==2.2.2
nltk==3.8.1
certifi==2023.5.7
gensim==4.3.1
pandas==2.0.2
tqdm==4.65.0
transformers==4.30.2
torch==2.0.1
readability-lxml==0.8.1
logging==0.5.1.2  
openai==0.27.8
spacy==3.2.0
